Serial SD improved, Gvar editing on expo, higher resolution switch timers

+++ src/simulatordialog.h
@@ -48,7 +48,7 @@ private:
     quint8  bpanaCenter;
     bool    swOn[MAX_MIXERS];
     quint16 one_sec_precount;
-		qint8		CsTimer[NUM_CSW] ;
+		qint16		CsTimer[NUM_CSW] ;
 
     quint16 s_timeCumTot;
     quint16 s_timeCumAbs;
